Add support for piping to stdout
Not sure if this is actually the desired behavior, but this
will allow for things like:

  generate-json | jpterm | process-result-json

Once you find an expression you like you can just hit
"ctrl r" and once you exit it will pipe the result to the
next command.

Another option is just to support writing the output of the command
to a file.